Add more advanced url exclusion testing
**Overview**
Previously the code took an array of strings for the excludeUrls option, specified in JSON, and tested a request's URL via Array.prototype.indexOf(). This needed updating. In order to do this the new code now takes several different options

0. We take an actual regular expression object if the options are defined outside of a regular expression
0. We take an array of two elements for each exclusion URL. The first is the regular expression string (in escaped string format) and the second is a string for options (such as "i" or "g").
0. Finally we take a string. The string is now first tested as regular expression and secondly as an Array.indexOf() match used in the older code.

**Finally**
The code previously ignored CSRF testing for some of the non-mutating
REST methods, but this has been modified to now include GET, HEAD,
OPTIONS, and TRACE.